requirements.txt
Pillow==9.5.0
PyPDF2==3.0.1
streamlit==1.23.1
transformers==4.30.2
